+++ components/vf-analytics-google/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.0.0-rc.3
+
+* documentation cleanup
+
 ### 1.0.0-rc.2
 
 * more robust fallback handling of image-based links